-- refatoring -- removed fixed variables -- create file paths -- added csv format
