+++ config.properties
@@ -1,4 +1,11 @@
 user=
 token=
 base_url=
+api_version=api-version=6.0
 base_path=
+############## excludes configs
+branch_prefix_exclude=TEST,development
+repo_prefix_exclude=TEST,development
+
+customer_branch_path=sim3g.cliente.${customer}.web
+branch_file_name=Branch
