+++ start.bat
@@ -1,1 +1,2 @@
-node BranchProcess.js planilha
+npm start xlsx
+pause
